lenny-podcast-analyzer/requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
pandas
python-dotenv
